+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog v1.1.0][keep-a-changelog]. See [the REA
 - Haskell SDK and Haskell module configuration. The project configuration is now determined automatically, no manual SDK and module setup required.
 - Formatting using stylish-haskell: since we are moving towards HLS anyway, there are no plans to support installation of tools not supported by the older LTS.
 - Settings for **Template name for new project** and **Cache path**.
+- Separate utilities for Ormolu formatting — we now rely on an extension for the default formatting service.
 
 ### Added
 - Compatibility with many other IntelliJ-based IDEs.